requirements.txt
flask
flask-cors
newspaper3k
transformers
torch
lxml_html_clean
